Correctly classifying known segment types in HCP mode
Former-commit-id: 5678db50c61355ad9762c582e6038e2740a51de7

+++ paradis.h
@@ -700,7 +700,7 @@ namespace paraDIS {
      /*!
       The burgers-type is defined above.  
     */ 
-    int8_t mBurgersType, mOriginalBurgersType; 
+    int16_t mBurgersType, mOriginalBurgersType; 
 
      /*!
       The screw type is defined above and is derived from the burgers type.
